Split websocket worker into its own module
This way all the base app is doing is loading the web server and passing
off to the different components to actually register and handle their
work. This should make it simpler to test as well.
