+++ index.js
@@ -1,95 +1,14 @@
 #!/usr/bin/env node
 
-const CIDRMatcher = require('cidr-matcher');
-const http = require('http');
-const net = require('net');
-const querystring = require('querystring');
-const sockjs = require('sockjs');
+'use strict'
 
 const config = require('./config');
-
-var matcher = new CIDRMatcher([ '10.0.0.0/8', '172.16.30.0/24' ]);
-
-var sockjs_server = sockjs.createServer({
-    sockjs_url: "https://cdn.jsdelivr.net/sockjs/1.1.2/sockjs.min.js"
-});
-
-sockjs_server.on('connection', function(conn) {
-
-    loc = conn.url.indexOf("?");
-
-    if (loc < 0) {
-        config.logger.app.info("no query string found");
-        conn.close()
-        return;
-    }
-
-    qs = conn.url.substring(loc + 1)
-    params = querystring.parse(qs);
-    host = params["host"];
-
-    if (!host) {
-        config.logger.app.info('host not specified in query string: ' + qs);
-        conn.close()
-        return;
-    }
-
-    if (!matcher.contains(host)) {
-        config.logger.app.warning("Attempt to proxy to an unauthorized IP: " + host);
-        conn.close();
-        return;
-    }
-
-    var client = net.connect(19885, host, function() {
-        // this callback gets triggered when a successful connection is established
-
-        client.on('close', function(had_error) {
-            config.logger.telnet.debug("telnet connection closed");
-            conn.close();
-        });
-
-        client.on('data', function(data) {
-            conn.write(data.toString());
-        });
-
-        // client.on('connect', function() {});
-
-        // client.on('drain', function() {});
-        // client.on('end', function() {});
-
-        // client.on('timeout', function() {});
-    });
-
-    client.on('error', function(err) {
-        // this handler is outside the connect callback to handle errors before
-        // connect occurs
-        config.logger.telnet.warning("Error: " + err.message);
-        conn.write('error: ' +  err.message);
-        conn.end()
-    });
-
-    config.logger.app.debug("setup complete to IP: " + host);
-
-    conn.on('data', function(message) {
-        // we shouldn't receive any information from the websocket.
-        config.logger.telnet.warning("data received on websocket: " + message);
-    });
-
-    conn.on('close', function() {
-        client.end();
-    });
-
-});
+const http = require('http');
 
 var server = http.createServer();
 
 require('./static')(server);
-
-server.addListener('upgrade', function(req,res) {
-    res.end();
-});
-
-sockjs_server.installHandlers(server, { prefix: '/sock' });
+require('./sock')(server);
 
 config.logger.app.info('listening on ' + config.bind_host + ':' + config.port);
 server.listen(config.port, config.bind_host);
